Unit tests on VivLib for VIV file data persistance
